Broke apart tf and k8s manifests
Merged Thomas' tf stuff into the environments module
Moved Shah's secret stuff into util as it's now shared between k8s and tf
Moved all the non k8s stuff into the terraform templates
