+++ templates/kubernetes/terraform/environments/staging/main.tf
@@ -1,32 +1,37 @@
 terraform {
   backend "s3" {
     bucket         = "project-{{ .Config.Name }}-terraform-state"
-    key            = "infrastructure/terraform/environments/staging/main"
+    key            = "infrastructure/terraform/environments/staging/kubernetes"
     encrypt        = true
     region         = "{{ .Config.Infrastructure.AWS.Region }}"
     dynamodb_table = "{{ .Config.Name }}-terraform-state-locks"
   }
 }
 
-# Instantiate the staging environment
-module "staging" {
-  source      = "../../modules/environment"
+# Provision kubernetes resources required to run services/applications
+module "kubernetes" {
+  source = "../../modules/kubernetes"
+
   environment = "staging"
+  region      = "{{ .Config.Infrastructure.AWS.Region }}"
 
-  # Project configuration
-  project             = "{{ .Config.Infrastructure.AWS.EKS.ClusterName }}"
-  region              = "{{ .Config.Infrastructure.AWS.Region }}"
-  allowed_account_ids = ["{{ .Config.Infrastructure.AWS.AccountId }}"]
+  # Authenticate with the EKS cluster via the cluster id
+  cluster_name = "{{ .Config.Infrastructure.AWS.EKS.ClusterName }}"
 
-  # ECR configuration
-  ecr_repositories = ["{{ .Config.Infrastructure.AWS.EKS.ClusterName }}"]
+  # Assume-role policy used by monitoring fluentd daemonset
+  assume_role_policy = data.aws_iam_policy_document.assumerole_root_policy.json
+}
 
-  # EKS configuration
-  eks_worker_instance_type = "t2.small"
-  eks_worker_asg_max_size  = 2
+# Data sources for EKS IAM
+data "aws_caller_identity" "current" {}
 
-  # EKS-Optimized AMI for your region: https://docs.aws.amazon.com/eks/latest/userguide/eks-optimized-ami.html
-  # https://us-east-1.console.aws.amazon.com/systems-manager/parameters/%252Faws%252Fservice%252Feks%252Foptimized-ami%252F1.14%252Famazon-linux-2%252Frecommended%252Fimage_id/description?region=us-east-1
-  eks_worker_ami = "{{ .Config.Infrastructure.AWS.EKS.WorkerAMI }}"
+data "aws_iam_policy_document" "assumerole_root_policy" {
+  statement {
+    actions = ["sts:AssumeRole"]
 
+    principals {
+      type        = "AWS"
+      identifiers = ["arn:aws:iam::${data.aws_caller_identity.current.account_id}:root"]
+    }
+  }
 }
