--- templates/kubernetes/terraform/modules/vpc/versions.tf
@@ -1,4 +1,0 @@
-
-terraform {
-  required_version = ">= 0.12"
-}
