+++ templates/terraform/bootstrap/remote-state/main.tf
