+++ templates/terraform/environments/development/main.tf
@@ -1,4 +1,5 @@
 terraform {
+  required_version = ">= 0.12"
   backend "s3" {
     bucket         = "project-{{ .Config.Name }}-terraform-state"
     key            = "infrastructure/terraform/environments/development/main"
@@ -10,7 +11,7 @@ terraform {
 
 # Instantiate the development environment
 module "development" {
-  source      = "../../../modules/environment"
+  source      = "../../modules/environment"
   environment = "development"
 
   # Project configuration
@@ -18,6 +19,7 @@ module "development" {
   region              = "{{ .Config.Infrastructure.AWS.Region }}"
   allowed_account_ids = ["{{ .Config.Infrastructure.AWS.AccountId }}"]
 
+  {{- if ne .Config.Infrastructure.AWS.EKS.ClusterName "" }}
   # ECR configuration
   ecr_repositories = ["{{ .Config.Infrastructure.AWS.EKS.ClusterName }}"]
 
@@ -28,5 +30,6 @@ module "development" {
   # EKS-Optimized AMI for your region: https://docs.aws.amazon.com/eks/latest/userguide/eks-optimized-ami.html
   # https://us-east-1.console.aws.amazon.com/systems-manager/parameters/%252Faws%252Fservice%252Feks%252Foptimized-ami%252F1.14%252Famazon-linux-2%252Frecommended%252Fimage_id/description?region=us-east-1
   eks_worker_ami = "{{ .Config.Infrastructure.AWS.EKS.WorkerAMI }}"
+  {{- end }}
 
 }
