--- templates/terraform/global/remote-state/main.tf
@@ -1,4 +1,0 @@
-provider "aws" {
-  region  = "${ var.region }"
-}
-
