--- templates/terraform/global/remote-state/terraform.tfvars
@@ -1,1 +1,0 @@
-region = "{{ .Config.Infrastructure.AWS.Region }}"
