+++ templates/terraform/modules/vpc/main.tf
