+++ templates/terraform/modules/vpc/outputs.tf
