+++ templates/terraform/modules/vpc/variables.tf
